Documents/bitspilani/bitsparinati/requirements.txt
amqp==2.1.4
anyjson==0.3.3
appdirs==1.4.0
awsebcli==3.9.1
Babel==2.3.3
backports-abc==0.4
backports.shutil-get-terminal-size==1.0.0
backports.ssl-match-hostname==3.5.0.1
billiard==3.5.0.2
blessed==1.14.1
boto==2.47.0
boto3==1.4.4
botocore==1.5.10
celery==4.0.2
cement==2.8.2
certifi==2021.10.8
cffi==1.9.1
colorama==0.3.7
configparser==3.5.0
cssselect==0.9.2
csvImporter==0.1.3.2
decorator==4.0.10
diff-match-patch==20110725.1
disposable-email-domains==0.0.5
Django==1.9.5
django-appconf==1.0.3
django-bootstrap-dynamic-formsets==0.4.5
django-bootstrap3==7.0.1
django-celery==3.2.1
django-celery-beat==1.0.1
django-celery-results==1.0.1
django-ckeditor==5.0.3
django-countries==3.4.1
django-crispy-forms==1.6.0
django-crudbuilder==0.2.4
django-datatable==0.2.0
django-daterange-filter==1.3.0
django-datetime-widget==0.9.3
django-digg-paginator==0.1.3
django-easy-pdf==0.1.1
django-excel==0.0.4
django-excel-response==2.0.1
django-extensions==1.9.1
django-eztables==0.3.2
django-filter==0.13.0
django-formset-js==0.5.0
django-import-export==0.4.5
django-jquery-js==2.1.4
django-jsonfield==1.0.1
django-localflavor==1.2
django-mysql==1.1.0
django-paginated-modelformset==0.0.2
django-paytm==0.1
django-phonenumber-field==1.1.0
django-pure-pagination==0.3.0
django-queryset-csv==0.3.3
django-registration==2.0.4
django-registration-redux==1.4
django-smtp-ssl==1.0
django-staticfiles==1.2.1
django-storages==1.6.3
django-suit==0.2.18
django-tables2==1.1.6
django-validated-file==2.0.1
django-widget-tweaks==1.4.1
django.js==0.8.1
djangorestframework==3.3.3
djangorestframework-bulk==0.2.1
dnspython==1.15.0
docker-py==1.7.2
dockerpty==0.4.1
docopt==0.6.2
docutils==0.13.1
easy-thumbnails==2.3
entrypoints==0.2.2
enum34==1.1.6
et-xmlfile==1.0.1
functools32==3.2.3.post2
futures==3.1.1
google-api-python-client==1.6.2
gunicorn==19.4.5
html5lib==1.0.1
httplib2==0.10.3
idna==2.2
ipaddress==1.0.18
ipykernel==4.3.1
ipython==4.2.1
ipython-genutils==0.1.0
ipywidgets==5.1.5
jdcal==1.2
Jinja2==2.8
jmespath==0.9.1
jsonfield==1.0.3
jsonschema==2.5.1
jupyter==1.0.0
jupyter-client==4.3.0
jupyter-console==5.0.0
jupyter-core==4.1.0
kombu==4.0.2
lxml==3.7.2
Markdown==2.6.6
MarkupSafe==0.23
mistune==0.7.3
msgpack-python==0.4.8
MySQL-python==1.2.5
nbconvert==4.2.0
nbformat==4.0.1
notebook==4.2.1
oauth2client==4.0.0
odfpy==0.9.6
olefile==0.44
openpyxl==2.3.5
packaging==16.8
params==0.1.4
pathlib2==2.1.0
pathspec==0.5.0
pexpect==4.2.1
phonenumberslite==8.10.19
pickleshare==0.7.4
Pillow==3.2.0
pip-review==0.5.2
prompt-toolkit==1.0.9
psycopg2==2.7.1
ptyprocess==0.5.1
pyasn1==0.2.2
pyasn1-modules==0.0.8
pycparser==2.17
pycrypto==2.6.1
pydns==2.3.6
PyDrive==1.3.1
pyexcel==0.2.4
pyexcel-io==0.2.1
pyexcel-ods==0.2.0
pyexcel-webio==0.0.7
pyexcel-xls==0.2.1
pyexcel-xlsx==0.2.1
Pygments==2.1.3
pyparsing==2.1.10
PyPDF2==1.25.1
pyquery==1.2.13
python-dateutil==2.6.0
python-magic==0.4.12
pytz==2016.10
PyYAML==3.12
pyzmq==15.3.0
qtconsole==4.2.1
redis==2.10.5
reportlab==3.5.23
requests==2.9.1
rsa==3.4.2
s3transfer==0.1.10
semantic-version==2.5.0
simplegeneric==0.8.1
singledispatch==3.4.0.3
six==1.10.0
sorl-thumbnail==12.3
tablib==0.11.2
tabulate==0.7.5
terminado==0.6
testtable==1.3.0
texttable==0.8.4
tornado==4.4b1
traitlets==4.3.1
uritemplate==3.0.0
validate-email==1.3
validators==0.11.1
vine==1.1.3
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.40.0
Werkzeug==0.12.2
widgetsnbextension==1.2.3
xhtml2pdf==0.2.3
xlrd==0.9.4
xlwt==1.2.0
django-celery-email==2.0.2
